--- __init__.py
@@ -1,7 +1,0 @@
-"""
-TopoZero: Minimal MCTS-driven analog topology generator using an Euler-trail view.
-Flat layout: vocab+adjacency, MCTS+state, SPICE I/O, and a tiny runner.
-"""
-
-__all__ = ["vocab_and_adjacency", "euler_mcts", "netlist_io"]
-
