+++ topogenie/__init__.py
@@ -1,0 +1,10 @@
+"""
+TopoGenie: Minimal MCTS-driven analog topology generator using an Euler-trail view.
+"""
+
+from . import vocab
+from .euler_mcts import EulerState, MCTS
+from .netlist_io import score_circuit
+
+__all__ = ["EulerState", "MCTS", "score_circuit", "vocab"]
+
